requirements.txt
django
django-allauth
django-bootstrap3
django-filter
django-invitations